+++ Editor/ScriptableBuildPipelineVersion.cs
@@ -1,8 +1,0 @@
-namespace UnityEditor.Build.Pipeline
-{
-    internal class ScriptableBuildPipelineVersion
-    {
-        public const string kPackageName = "com.unity.scriptablebuildpipeline";
-        public const string kPackageVersion = "1.10.0";
-    }
-}