+++ Editor/WriteTypes/SceneRawWriteOperation.cs
@@ -10,6 +10,7 @@ namespace UnityEditor.Build.Pipeline.WriteTypes
     /// Explicit implementation for writing a scene serialized file that can be used with the upcoming raw loading systems.
     /// </summary>
     [Serializable]
+    [Obsolete("SceneRawWriteOperation obsolete and no longer maintained.")]
     public class SceneRawWriteOperation : IWriteOperation
     {
         /// <inheritdoc />