+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.11.1] - 2020-08-11
+- Exposed the ScriptableBuildPipeline static class to allow setting per project properties from script.
+- Fixed an edge case where pruning the build cache would not run in Unity's batchmode.
+- Added Cache Server Config options to the Scriptable Build Pipeline UI.
+
 ## [1.10.0] - 2020-07-28
 - Added IBundleBuildParameters.ContiguousBundles option, which when enabled will improve asset loading times.
   - In testing, performance improvements varied from 10% improvement over all, with improvements up to 50% for large complex assets such as extensive UI prefabs.